+++ rust/feature-flags/src/flags/flag_analytics.rs
@@ -2,8 +2,8 @@ use anyhow::Result;
 use std::sync::Arc;
 use std::time::{SystemTime, UNIX_EPOCH};
 
-use crate::flag_request::FlagRequestType;
-use crate::redis::{Client as RedisClient, CustomRedisError};
+use crate::client::redis::{Client as RedisClient, CustomRedisError};
+use crate::flags::flag_request::FlagRequestType;
 
 const CACHE_BUCKET_SIZE: u64 = 60 * 2; // duration in seconds
 
@@ -37,7 +37,7 @@ pub async fn increment_request_count(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::test_utils::setup_redis_client;
+    use crate::utils::test_utils::setup_redis_client;
 
     #[tokio::test]
     async fn test_get_team_request_key() {
