+++ rust/feature-flags/src/flags/flag_operations.rs
@@ -1,51 +1,11 @@
-use crate::{
-    api::FlagError, cohort_models::CohortId, database::Client as DatabaseClient,
-    redis::Client as RedisClient,
-};
-use serde::{Deserialize, Serialize};
+use crate::api::errors::FlagError;
+use crate::client::database::Client as DatabaseClient;
+use crate::client::redis::Client as RedisClient;
+use crate::cohort::cohort_models::CohortId;
+use crate::flags::flag_models::*;
+use crate::properties::property_models::PropertyFilter;
 use std::sync::Arc;
 use tracing::instrument;
-
-// TRICKY: This cache data is coming from django-redis. If it ever goes out of sync, we'll bork.
-// TODO: Add integration tests across repos to ensure this doesn't happen.
-pub const TEAM_FLAGS_CACHE_PREFIX: &str = "posthog:1:team_feature_flags_";
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum OperatorType {
-    Exact,
-    IsNot,
-    Icontains,
-    NotIcontains,
-    Regex,
-    NotRegex,
-    Gt,
-    Lt,
-    Gte,
-    Lte,
-    IsSet,
-    IsNotSet,
-    IsDateExact,
-    IsDateAfter,
-    IsDateBefore,
-    In,
-    NotIn,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct PropertyFilter {
-    pub key: String,
-    // TODO: Probably need a default for value?
-    // incase operators like is_set, is_not_set are used
-    // not guaranteed to have a value, if say created via api
-    pub value: serde_json::Value,
-    pub operator: Option<OperatorType>,
-    #[serde(rename = "type")]
-    // TODO: worth making a enum here to differentiate between cohort and person filters?
-    pub prop_type: String,
-    pub negation: Option<bool>,
-    pub group_type_index: Option<i32>,
-}
 
 impl PropertyFilter {
     /// Checks if the filter is a cohort filter
@@ -61,64 +21,6 @@ impl PropertyFilter {
         }
         self.value.as_i64().map(|id| id as CohortId)
     }
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct FlagGroupType {
-    pub properties: Option<Vec<PropertyFilter>>,
-    pub rollout_percentage: Option<f64>,
-    pub variant: Option<String>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct MultivariateFlagVariant {
-    pub key: String,
-    pub name: Option<String>,
-    pub rollout_percentage: f64,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct MultivariateFlagOptions {
-    pub variants: Vec<MultivariateFlagVariant>,
-}
-
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct FlagFilters {
-    pub groups: Vec<FlagGroupType>,
-    pub multivariate: Option<MultivariateFlagOptions>,
-    pub aggregation_group_type_index: Option<i32>,
-    pub payloads: Option<serde_json::Value>,
-    pub super_groups: Option<Vec<FlagGroupType>>,
-}
-
-// TODO: see if you can combine these two structs, like we do with cohort models
-// this will require not deserializing on read and instead doing it lazily, on-demand
-// (which, tbh, is probably a better idea)
-#[derive(Debug, Clone, Deserialize, Serialize)]
-pub struct FeatureFlag {
-    pub id: i32,
-    pub team_id: i32,
-    pub name: Option<String>,
-    pub key: String,
-    pub filters: FlagFilters,
-    #[serde(default)]
-    pub deleted: bool,
-    #[serde(default)]
-    pub active: bool,
-    #[serde(default)]
-    pub ensure_experience_continuity: bool,
-}
-
-#[derive(Debug, Serialize, sqlx::FromRow)]
-pub struct FeatureFlagRow {
-    pub id: i32,
-    pub team_id: i32,
-    pub name: Option<String>,
-    pub key: String,
-    pub filters: serde_json::Value,
-    pub deleted: bool,
-    pub active: bool,
-    pub ensure_experience_continuity: bool,
 }
 
 impl FeatureFlag {
@@ -144,11 +46,6 @@ impl FeatureFlag {
                 .and_then(|obj| obj.get(match_val).cloned())
         })
     }
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct FeatureFlagList {
-    pub flags: Vec<FeatureFlag>,
 }
 
 impl FeatureFlagList {
@@ -243,14 +140,14 @@ impl FeatureFlagList {
 
 #[cfg(test)]
 mod tests {
-    use crate::flag_definitions;
+    use crate::{flags::flag_models::*, properties::property_models::OperatorType};
     use rand::Rng;
     use serde_json::json;
     use std::time::Instant;
     use tokio::task;
 
     use super::*;
-    use crate::test_utils::{
+    use crate::utils::test_utils::{
         insert_flag_for_team_in_pg, insert_flags_for_team_in_redis, insert_new_team_in_pg,
         insert_new_team_in_redis, setup_invalid_pg_client, setup_pg_reader_client,
         setup_redis_client,
@@ -803,6 +700,7 @@ mod tests {
             }
         }
     }
+
     #[tokio::test]
     async fn test_flag_with_super_groups() {
         let redis_client = setup_redis_client(None);
@@ -1114,7 +1012,7 @@ mod tests {
 
         redis_client
             .set(
-                format!("{}{}", flag_definitions::TEAM_FLAGS_CACHE_PREFIX, team.id),
+                format!("{}{}", TEAM_FLAGS_CACHE_PREFIX, team.id),
                 "not a json".to_string(),
             )
             .await
@@ -1300,7 +1198,7 @@ mod tests {
                 "id": 2,
                 "team_id": team.id,
                 "name": "Very Long Key Flag",
-                "key": "a".repeat(400), // max key length is 400
+                "key": "aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "filters": {"groups": [{"properties": [], "rollout_percentage": 100}]},
                 "active": true,
                 "deleted": false
