+++ rust/feature-flags/tests/test_flags.rs
@@ -7,7 +7,7 @@ use serde_json::{json, Value};
 use crate::common::*;
 
 use feature_flags::config::DEFAULT_TEST_CONFIG;
-use feature_flags::test_utils::{
+use feature_flags::utils::test_utils::{
     insert_flags_for_team_in_redis, insert_new_team_in_pg, insert_new_team_in_redis,
     setup_pg_reader_client, setup_redis_client,
 };
